Add files via upload
The Love Calculator work out the love score between two people:

it takes both people's names and check for the number of times the letters in the word TRUE occurs. 

Then check for the number of times the letters in the word LOVE occurs. 

Then combines these numbers to make a 2 digit number.